Newdig: Placing doors.
