+++ keeperfx/keeperfx_vs2010.vcxproj.filters
@@ -681,6 +681,9 @@
     <ClCompile Include="src\player_newcomppathfind.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="src\player_newcompdigging.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\bflib_base_tcp.hpp">
